fastapi_solution/src/requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
fastapi==0.100.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
orjson==3.9.10
pydantic==2.5.1
pydantic_core==2.14.3
pydantic_settings==2.1.0
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
gunicorn==20.0.4
uvicorn==0.24.0.post1
yarl==1.9.2
aiohttp==3.9.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
async-fastapi-jwt-auth==0.6.2
python-jose[cryptography]==3.3.0
sentry-sdk[fastapi]